app/requirements.txt
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
ecdsa==0.14.1
fastapi==0.61.1
h11==0.10.0
neo4j==4.1.1
numpy==1.19.2
pandas==1.1.2
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
rsa==4.6
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.1
